+++ CHANGELOG.md
@@ -5,6 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v0.3.0] - 2023-12-21
+
+### Added
+
+- `resource_project` includes two additional attributes to configure IP addresses allowed to connect to the project's 
+endpoints:
+  - `allowed_ips`
+  - `allowed_ips_primary_branch_only`
+
+### Fixed
+
+- Schema is set on per resource basis now.
+
+### Changed
+
+- Updated dependencies: Neon Go SDK [v0.4.0](https://github.com/kislerdm/neon-sdk-go/releases/tag/v0.4.0)
+
 ## [v0.2.5] - 2023-11-02
 
 ### Fixed